+++ R/prediction_biglm.R
@@ -17,7 +17,6 @@ function(model,
     } else {
         # reduce memory profile
         model[["model"]] <- NULL
-        attr(model[["terms"]], ".Environment") <- NULL
         
         # setup data
         data <- build_datalist(data, at = at, as.data.frame = TRUE)
@@ -34,11 +33,18 @@ function(model,
         }
     }
     
-    # obs-x-(ncol(data)+2) data frame
+    # variance(s) of average predictions
+    vc <- NA_real_
+    
+    # output
     structure(pred, 
               class = c("prediction", "data.frame"),
-              row.names = seq_len(nrow(pred)),
               at = if (is.null(at)) at else at_specification,
-              model.class = class(model),
-              type = type)
+              type = type,
+              call = if ("call" %in% names(model)) model[["call"]] else NULL,
+              model_class = class(model),
+              row.names = seq_len(nrow(pred)),
+              vcov = vc,
+              jacobian = NULL,
+              weighted = FALSE)
 }