+++ R/prediction_gausspr.R
@@ -9,6 +9,8 @@ function(model,
          category, 
          ...) {
     
+    requireNamespace("kernlab")
+    
     if (!is.null(type)) {
         warning(sprintf("'type' is ignored for models of class '%s'", class(model)))
     }
@@ -16,8 +18,8 @@ function(model,
     # extract predicted values
     data <- data
     if (missing(data) || is.null(data)) {
-        pred <- make_data_frame(fitted.class = predict(model, type = "response", ...))
-        probs <- make_data_frame(predict(model, type = "probabilities", ...))
+        pred <- make_data_frame(fitted.class = kernlab::predict(model, type = "response", ...))
+        probs <- make_data_frame(kernlab::predict(model, type = "probabilities", ...))
         names(probs) <- paste0("Pr(", names(probs), ")")
         pred <- cbind(pred, probs)
     } else {
@@ -29,8 +31,8 @@ function(model,
             at_specification <- attr(out, "at_specification")
         }
         # calculate predictions
-        tmp <- predict(model, newdata = out, type = "response", ...)
-        tmp_probs <- make_data_frame(predict(model, newdata = out, type = "probabilities", ...))
+        tmp <- kernlab::predict(model, newdata = out, type = "response", ...)
+        tmp_probs <- make_data_frame(kernlab::predict(model, newdata = out, type = "probabilities", ...))
         names(tmp_probs) <- paste0("Pr(", names(tmp_probs), ")")
         # cbind back together
         pred <- make_data_frame(out, fitted.class = tmp, tmp_probs)
@@ -51,12 +53,19 @@ function(model,
     }
     pred[["se.fitted"]] <- NA_real_
     
-    # obs-x-(ncol(data)+2+nlevels(outcome)) data.frame of predictions
-    structure(pred,
-              class = c("prediction", "data.frame"), 
-              row.names = seq_len(nrow(pred)),
+    # variance(s) of average predictions
+    vc <- NA_real_
+    
+    # output
+    structure(pred, 
+              class = c("prediction", "data.frame"),
               at = if (is.null(at)) at else at_specification,
-              model.class = class(model),
               type = NA_character_,
-              category = category)
+              call = if ("call" %in% names(model)) model[["call"]] else NULL,
+              model_class = class(model),
+              row.names = seq_len(nrow(pred)),
+              vcov = vc,
+              jacobian = NULL,
+              category = category,
+              weighted = FALSE)
 }