+++ R/prediction_kqr.R
@@ -2,9 +2,11 @@
 #' @export
 prediction.kqr <- function(model, data, at = NULL, calculate_se = FALSE, ...) {
     
+    requireNamespace("kernlab")
+    
     # extract predicted values
     if (missing(data) || is.null(data)) {
-        pred <- make_data_frame(fitted = predict(object = model, ...)[,1L])
+        pred <- make_data_frame(fitted = kernlab::predict(object = model, ...)[,1L])
     } else {
         # setup data
         if (is.null(at)) {
@@ -13,15 +15,22 @@ prediction.kqr <- function(model, data, at = NULL, calculate_se = FALSE, ...) {
             out <- build_datalist(data, at = at, as.data.frame = TRUE)
             at_specification <- attr(out, "at_specification")
         }
-        pred <- make_data_frame(fitted = predict(model, newdata = data,...)[,1L])
+        pred <- make_data_frame(fitted = kernlab::predict(model, newdata = data,...)[,1L])
     }
     pred[["se.fitted"]] <- NA_real_
     
-    # obs-x-(ncol(data)+2) data frame
+    # variance(s) of average predictions
+    vc <- NA_real_
+    
+    # output
     structure(pred, 
-              class = c("prediction", "data.frame"), 
-              row.names = seq_len(nrow(pred)),
+              class = c("prediction", "data.frame"),
               at = if (is.null(at)) at else at_specification,
-              model.class = class(model),
-              type = NA_character_)
+              type = NA_character_,
+              call = if ("call" %in% names(model)) model[["call"]] else NULL,
+              model_class = class(model),
+              row.names = seq_len(nrow(pred)),
+              vcov = vc,
+              jacobian = NULL,
+              weighted = FALSE)
 }