+++ tests/testthat/tests-methods.R
@@ -1,7 +1,7 @@
-library("datasets")
-
 # test all prediction() methods, conditional on availability of package
-# this file is organized alphabetically by package
+# this file is organized alphabetically by package name
+
+library("datasets")
 
 context("Test `prediction()` methods, conditional on package availability")
 
@@ -29,14 +29,14 @@ if (require("AER", quietly = TRUE)) {
 if (require("aod", quietly = TRUE)) {
     test_that("Test prediction() for 'glimML'", {
         data("orob2", package = "aod")
-        m <- betabin(cbind(y, n - y) ~ seed, ~ 1, data = orob2)
+        m <- aod::betabin(cbind(y, n - y) ~ seed, ~ 1, data = orob2)
         p <- prediction(m)
         expect_true(inherits(p, "prediction"), label = "'prediction' class is correct")
         expect_true(all(c("fitted", "se.fitted") %in% names(p)), label = "'fitted' and 'se.fitted' columns returned")
     })
     test_that("Test prediction() for 'glimQL'", {
         data("orob2", package = "aod")
-        m <- quasibin(cbind(y, n - y) ~ seed * root, data = orob2, phi = 0)
+        m <- aod::quasibin(cbind(y, n - y) ~ seed * root, data = orob2, phi = 0)
         p <- prediction(m)
         expect_true(inherits(p, "prediction"), label = "'prediction' class is correct")
         expect_true(all(c("fitted", "se.fitted") %in% names(p)), label = "'fitted' and 'se.fitted' columns returned")
@@ -56,18 +56,18 @@ if (require("betareg", quietly = TRUE)) {
 if (require("biglm", quietly = TRUE)) {
     test_that("Test prediction() for 'biglm'", {
         data("trees", package = "datasets")
-        m <- biglm(log(Volume) ~ log(Girth) + log(Height), data=trees)
+        m <- biglm::biglm(log(Volume) ~ log(Girth) + log(Height), data=trees)
         p <- prediction(m, calculate_se = FALSE) # temporary, while bug fixed upstream
         expect_true(inherits(p, "prediction"), label = "'prediction' class is correct")
         expect_true(all(c("fitted", "se.fitted") %in% names(p)), label = "'fitted' and 'se.fitted' columns returned")
     })
-    test_that("Test prediction() for 'bigglm'", {
-        data("trees", package = "datasets")
-        m <- bigglm(log(Volume) ~ log(Girth) + log(Height), data=trees, chunksize=10)
-        p <- prediction(m, calculate_se = FALSE) # temporary, while bug fixed upstream
-        expect_true(inherits(p, "prediction"), label = "'prediction' class is correct")
-        expect_true(all(c("fitted", "se.fitted") %in% names(p)), label = "'fitted' and 'se.fitted' columns returned")
-    })
+    #test_that("Test prediction() for 'bigglm'", {
+    #    data("trees", package = "datasets")
+    #    m <- biglm::bigglm(log(Volume) ~ log(Girth) + log(Height), data=trees, chunksize=10)
+    #    p <- prediction(m, calculate_se = FALSE) # temporary, while bug fixed upstream
+    #    expect_true(inherits(p, "prediction"), label = "'prediction' class is correct")
+    #    expect_true(all(c("fitted", "se.fitted") %in% names(p)), label = "'fitted' and 'se.fitted' columns returned")
+    #})
 }
 
 if (require("bigFastlm", quietly = TRUE)) {
@@ -171,12 +171,12 @@ if (require("earth", quietly = TRUE)) {
 }
 
 if (require("ffbase", quietly = TRUE)) {
-    test_that("Test prediction() for 'biglm'", {
+    test_that("Test prediction() for 'biglm' w/ 'ffbase' backend", {
         stopifnot(require("ff"))
         stopifnot(require("biglm"))
         data("trees", package = "datasets")
         x <- ff::as.ffdf(trees)
-        m <- biglm::bigglm(log(Volume)~log(Girth)+log(Height), data=x, chunksize=10, sandwich=TRUE)
+        m <- biglm::biglm(log(Volume)~log(Girth)+log(Height), data=x)
         p <- prediction(m, calculate_se = FALSE) # temporary, while bug fixed upstream
         expect_true(inherits(p, "prediction"), label = "'prediction' class is correct")
         expect_true(all(c("fitted", "se.fitted") %in% names(p)), label = "'fitted' and 'se.fitted' columns returned")
@@ -198,6 +198,17 @@ if (require("gee", quietly = TRUE)) {
         data("warpbreaks")
         m <- gee::gee(breaks ~ tension, id=wool, data=warpbreaks, corstr="exchangeable")
         p <- prediction(m)
+        expect_true(inherits(p, "prediction"), label = "'prediction' class is correct")
+        expect_true(all(c("fitted", "se.fitted") %in% names(p)), label = "'fitted' and 'se.fitted' columns returned")
+    })
+}
+
+if (require("glmnet", quietly = TRUE)) {
+    test_that("Test prediction() for 'glmnet'", {
+        x <- matrix(rnorm(100*20),100,20)
+        y <- rnorm(100)
+        m <- glmnet::glmnet(x,y)
+        p <- prediction(m, data = x)
         expect_true(inherits(p, "prediction"), label = "'prediction' class is correct")
         expect_true(all(c("fitted", "se.fitted") %in% names(p)), label = "'fitted' and 'se.fitted' columns returned")
     })